feat: Break up god class
